requirements.txt
django
Pillow
djangorestframework
psycopg2-binary
django-ckeditor
django-filter
djoser
djangorestframework_simplejwt
drf-yasg
django-rest-framework-social-oauth2
django-cors-headers